Bugfix merge from 0.96rc1 (this is now 0.96.1 dev)

+++ modules/sharedcoremodule/actions/activate_site.php
@@ -33,24 +33,45 @@
 
 if (!defined("PATHOS")) exit("");
 
-// PERM CHECK
-	$site = $db->selectObject("sharedcore_site","id=".$_GET['id']);
+if (pathos_permissions_check('manage_site',pathos_core_makeLocation('sharedcoremodule'))) {
+	$site = null;
+	if (isset($_GET['id'])) {
+		$site = $db->selectObject("sharedcore_site","id=".$_GET['id']);
+	}
+	
 	if ($site) {
 		$core = $db->selectObject("sharedcore_core","id=".$site->core_id);
 		if ($core) {
 			$site->inactive = 0;
 			$db->updateObject($site,"sharedcore_site");
+			
 			unlink($site->path."index.php");
 			
-			$extensions = $db->selectObjects("sharedcore_extension","site_id=".$site->id);
 			if (!defined("SYS_SHAREDCORE")) include_once(BASE."subsystems/sharedcore.php");
-			pathos_sharedcore_linkCore($core->path,$site->path);
-			foreach ($extensions as $e) {
-				pathos_sharedcore_linkExtension($e->type,$e->name,$core->path,$site->path);
+			
+			pathos_sharedcore_setup($core,$site);
+			
+			$extensions = array(
+				CORE_EXT_MODULE=>array(),
+				CORE_EXT_SUBSYSTEM=>array(),
+				CORE_EXT_THEME=>array(),
+			);
+			
+			foreach ($db->selectObjects("sharedcore_extension","site_id=".$site->id) as $e) {
+				$extensions[$e->type][] = $e->name;
 			}
+			
+			pathos_sharedcore_link($core,$site,$extensions);
+			
 			pathos_flow_redirect();
-		} else echo SITE_404_HTML; // temp
-	} else echo SITE_404_HTML;
-// END PERM CHECK
+		} else {
+			echo SITE_404_HTML;
+		}
+	} else {
+		echo SITE_404_HTML;
+	}
+} else {
+	echo SITE_403_HTML;
+}
 
 ?>
